@@ -1,25 +1,27 @@
+/*
 package vn.com.ibss.au.validation;
-
-import javax.validation.Constraint;
-import javax.validation.Payload;
+ 
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+ 
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
-
+ 
 import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target({TYPE, ANNOTATION_TYPE})
+ 
+@Target({ TYPE, ANNOTATION_TYPE })
 @Retention(RUNTIME)
 @Constraint(validatedBy = PasswordMatchesValidator.class)
 @Documented
 public @interface PasswordMatches {
-
+ 
     String message() default "Passwords don't match";
-
+ 
     Class<?>[] groups() default {};
-
+ 
     Class<? extends Payload>[] payload() default {};
-
-}
+ 
+}*/
