@@ -1,19 +1,14 @@
+/*
 package vn.com.ibss.au.validation;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-public class PasswordMatchesValidator implements ConstraintValidator<PasswordMatches, Object> {
-
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+ 
+public class PasswordMatchesValidator implements ConstraintValidator<PasswordMatches, SignUpRequest> {
+ 
     @Override
-    public void initialize(final PasswordMatches constraintAnnotation) {
+    public boolean isValid(final SignUpRequest user, final ConstraintValidatorContext context) {
+        return user.getPassword().equals(user.getMatchingPassword());
     }
-
-    @Override
-    public boolean isValid(final Object obj, final ConstraintValidatorContext context) {
-       // final UserDTO user = (UserDTO) obj;
-        //return user.getPassword().equals(user.getMatchingPassword());
-        return true;
-    }
-
-}
+ 
+}*/
